# solution/main_server/swagger/swaggerDocumentationMainServer.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Football Analyzer API",
    "description": "API documentation for the Football application. Reminder: If you want to try a particular route which requires two or more parameters, and you want to filter by one or two of them, just insert an empty string ('''') in the field you want to leave empty.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "main express server"
    },
    {
      "url": "http://localhost:4000",
      "description": "second express server"
    },
    {
      "url": "http://localhost:8080",
      "description": "springboot server"
    }
  ],
  "paths": {
...
    "/springboot/get_clubs": {
      "get": {
        "summary": "Get clubs from the springboot server",
        "description": "Fetches the top clubs",
        "responses": {
          "200": {
            "description": "A list clubs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpringbootClubs"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        },
        "tags": ["localhost:8080"]
      }
    },
    "/springboot/get_filtered_clubs": {
      "get": {
        "summary": "Get filtered clubs from the springboot server",
        "description": "Fetches clubs filtered by name, competition and country",
        "parameters": [
          {
            "in": "query",
            "name": "clubName",
            "schema": {
              "type": "string"
            },
            "description": "The name of the club"
          },
          {
            "in": "query",
            "name": "competition",
            "schema": {
              "type": "string"
            },
            "description": "The competition of the club"
          },
          {
            "in": "query",
            "name": "country",
            "schema": {
              "type": "string"
            },
            "description": "The country of the club"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string"
            },
            "description": "The limit for the number of results"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "string"
            },
            "description": "The offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of filtered clubs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpringbootFilteredClubs"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No clubs found matching the provided filters"
          },
          "500": {
            "description": "Server error"
          }
        },
        "tags": ["localhost:8080"]
      }
    },
    "/springboot/get_club_names": {
      "get": {
        "summary": "Get clubs names from the springboot server",
        "description": "Fetches club names by their IDs",
        "parameters": [
          {
            "in": "query",
            "name": "club_ids",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "description": "List of club IDs"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of club names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubNames"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        },
        "tags": ["localhost:8080"]
      }
    }
  },
  "components": {
    "schemas": {
...
      "SpringbootClubs": {
        "type": "object",
        "properties": {
          "club_id": {
            "type": "integer",
            "example": 31
          },
          "name": {
            "type": "string",
            "example": "Liverpool Football Club"
          },
          "domestic_competition_id": {
            "type": "string",
            "example": "GB1"
          },
          "domestic_competition_name": {
            "type": "string",
            "example": "premier-league"
          },
          "squad_size": {
            "type": "integer",
            "example": 26
          },
          "average_age": {
            "type": "number",
            "format": "float",
            "example": 26.2
          },
          "foreigners_number": {
            "type": "integer",
            "example": 21
          },
          "foreigners_percentage": {
            "type": "number",
            "format": "float",
            "example": 80.8
          },
          "national_team_players": {
            "type": "integer",
            "example": 17
          },
          "stadium_name": {
            "type": "string",
            "example": "Anfield"
          },
          "stadium_seats": {
            "type": "integer",
            "example": 54074
          },
          "net_transfer_record": {
            "type": "string",
            "example": "€-111.30m"
          },
          "last_season": {
            "type": "integer",
            "example": 2023
          }
        }
      },
      "SpringbootFilteredClubs": {
        "type": "object",
        "properties": {
          "club_id": {
            "type": "integer",
            "example": 506
          },
          "name": {
            "type": "string",
            "example": "Juventus Football Club"
          },
          "domestic_competition_id": {
            "type": "string",
            "example": "IT1"
          },
          "domestic_competition_name": {
            "type": "string",
            "example": "serie-a"
          },
          "squad_size": {
            "type": "integer",
            "example": 26
          },
          "average_age": {
            "type": "number",
            "format": "float",
            "example": 26.7
          },
          "foreigners_number": {
            "type": "integer",
            "example": 14
          },
          "foreigners_percentage": {
            "type": "number",
            "format": "float",
            "example": 53.8
          },
          "national_team_players": {
            "type": "integer",
            "example": 16
          },
          "stadium_name": {
            "type": "string",
            "example": "Allianz Stadium"
          },
          "stadium_seats": {
            "type": "integer",
            "example": 41507
          },
          "net_transfer_record": {
            "type": "string",
            "example": "€-21.60m"
          },
          "last_season": {
            "type": "integer",
            "example": 2023
          }
        }
      },
      "ClubNames": {
        "type": "object",
        "properties": {
          "club_id": {
            "type": "integer",
            "example": 506
          },
          "club_name": {
            "type": "string",
            "example": "Juventus Football Club"
          }
        }
      }
    }
  }
}